transfer constant data to mock data
